+++ crates/core/src/http/errors.rs
@@ -4,7 +4,7 @@ mod parse_error;
 cfg_feature! {
     #![feature = "rfc9457"]
     mod problem;
-    pub use problem::{PROBLEM_JSON, Problem};
+    pub use problem::{NoExtensions, PROBLEM_JSON, PlainProblem, Problem};
 }
 mod status_error;
 pub use parse_error::{ParseError, ParseResult};